+++ pkg/distributor/push_test.go
@@ -1539,6 +1539,8 @@ func (o otlpLimitsMock) OTelKeepIdentifyingResourceAttributes(string) bool {
 	return false
 }
 
+func (o otlpLimitsMock) OTelConvertHistogramsToNHCB(string) bool { return false }
+
 func promToMimirHistogram(h *prompb.Histogram) mimirpb.Histogram {
 	pSpans := make([]mimirpb.BucketSpan, 0, len(h.PositiveSpans))
 	for _, span := range h.PositiveSpans {